+++ acdc-lsp/src/convert.rs
@@ -54,12 +54,12 @@ pub(crate) fn offset_in_location(offset: usize, location: &Location) -> bool {
 pub(crate) fn location_to_range(loc: &Location) -> Range {
     Range {
         start: Position {
-            line: to_lsp_u32(loc.start.line.saturating_sub(1)),
-            character: to_lsp_u32(loc.start.column.saturating_sub(1)),
+            line: loc.start.line.saturating_sub(1),
+            character: loc.start.column.saturating_sub(1),
         },
         end: Position {
-            line: to_lsp_u32(loc.end.line.saturating_sub(1)),
-            character: to_lsp_u32(loc.end.column),
+            line: loc.end.line.saturating_sub(1),
+            character: loc.end.column,
         },
     }
 }